Removed init_app config saving for AutoPrefs in favor of dynamic manipulation.
Dynamically setting the PluginSwitch should be more than enough although this probably should be relegated to upgrade docs and a single upgrade action so that it doesn't have to do this on every initialization of the app in perpetuity.

Also, the second key (AutoPrefs) was added in case the plugin sig isn't used like it used to be.