Version 2.5.11: Fixed binary file corruption issue
static-copy was not using `binmode` on the file handle and hence was corrupting images.

However, it's difficult to do this with MT::FileMgr because it's geared for uploads and contains other logic for handling them. So, I switched to using File::Copy (bundled with perl) which significantly streamlined the code and works (bonus!)

Also simplified program output so that it's easier to see what's going on.