Version v2.5.23: Fixed "not a HASH reference" error
Using ref to figure out what kind of reference something is a bad practice, especially when you have array-based objects like MT::PluginData.  Since ref($plugindata_obj) yields 'MT::PluginData' and not ARRAY, the code was falling into the else block which assumes as HASH reference.

The correct way to check for the TYPE of reference is to either just use in a Try::Tiny try block OR, better, use Scalar::Util::reftype.